requirements.txt
# QAPal Dependencies

# Playwright for browser automation
playwright>=1.40.0

# TinyDB for locator storage
tinydb>=4.8.0

# AI Providers (install as needed)
anthropic>=0.18.0
openai>=1.0.0

# Optional utilities
python-dotenv>=1.0.0

# Semantic context extraction (optional — falls back to a11y snapshot if not installed)
crawl4ai>=0.3.0

# Visual regression diffing (optional — only needed when QAPAL_VISUAL_REGRESSION=true)
Pillow>=10.0.0

# HTTP client for non-standard AI endpoints (optional — only needed with custom QAPAL_AI_BASE_URL)
httpx>=0.25.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
